+++ pyspark_ds_toolbox/ml/__init__.py
@@ -4,4 +4,4 @@ Subpackage dedicated to Machine Learning helpers.
 """
 
 from pyspark_ds_toolbox.ml import data_prep, shap_values
-from pyspark_ds_toolbox.ml.classification import eval
+from pyspark_ds_toolbox.ml.classification import eval, baseline_classifiers