+++ pyspark_ds_toolbox/ml/classification/__init__.py
@@ -3,4 +3,4 @@ __doc__  = """Classification toolbox.
 Subpackage dedicated to Classification helpers.
 """
 
-from pyspark_ds_toolbox.ml.classification import eval
+from pyspark_ds_toolbox.ml.classification import eval, baseline_classifiers