+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### Added
 
-* 
+* Module pyspark_ds_toolbox.ml.classification
+
+### changed
+
+* Module pyspark_ds_toolbox.ml.eval became pyspark_ds_toolbox.ml.classification.eval
 
 ## 0.1.4